Ensure latest version of Hiera eYAML GPG
The Hiera eYAML GPG apt package is stored in our 'govuk' PPA Aptly
mirror, so we can control which version is available to each of our
environments.

Ensure that Puppet will always upgrade to the latest available
version, and make the version a class parameter in case we need to
override it in future.