Add `replicaset_members` parameter to server class
This is in preparation for having services call the server class with the members for the replicaset rather than calling the `configure_replica_set` class separately.

The parameter is an array which defaults to empty and throws an error if it's empty, i.e. you must specify the members.

Also add some tests. Do not yet use the configure replica set class via the server class so the tests pass but this won't actually work.