Fix pg_hba.conf rule for Bouncer
From testing with Vagrant it looked as though the slave would expect
authentication to appear to come from the backend gateway, but in fact
after testing in Staging this rule needs to allow authentication from
the Bouncer machines. Since nothing else is in the redirector vDC,
allow auth from any IP there.