Move graphite tarball creation from 0342 to 2010
The graphite tarball takes approximately 13 hours to create. If we start
at 0342, it runs until approximately 1432 which means for most of the
day our graphite server is tagged as having high IO.

By moving this to 2010 we give it a chance to complete before our
traffic starts ramping up for the morning (where we submit more
metrics), which has the advantage of possibly being faster to complete
(due to less general IO on the platform), but also not impacting metric
collection.