Improve pg_resync_slave downtime and disk usage
Use `mv` instead of `rsync` to put the new data into place after sync in
order to reduce the amount of time that the PosgreSQL service is stopped.
This is instantaneous when the source and destination are on the same
filesystem/mount.

This also prevents the old data from the sync being left on the filesystem
afterwards. Although you will still need to accommodate 2x the dataset size
during the sync.