pg_resync_slave don't error on server.{crt,key}
The `server.crt` and `server.key` symlinks don't exist on our new Ubuntu
Trusty machines which use the standard PostgreSQL 9.3 package, which causes
this script to fail. Prevent this from happening by catching the error. I've
separated this from the other move command, so it will still bail it's it
unable to copy `pg_xlog` or `recovery.tmp`, because they are essential.

It's not clear why these files don't exist on the new installs. PostgreSQL
9.1 on Precise puts them there and fails to start without them, but that
could either be a 9.1 thing or because we're using a custom (non-Ubuntu
upstream) package from Launchpad.