Clear out the `transition-postgres` app.
Now that we're going to make the Postgres version of the Transition app
the only version, we shouldn't keep this around.

Getting rid of the `Procfile` worker is the slightly tricky part, as
`govuk::procfile::worker` doesn't have a flag to clear out old workers.
A couple of `exec` resources will clear this up, in the same fashion as
used to clear up Imminence's old `delayed_job` workers in commit
4a8281540b9c9da085e07f7d2b1d268c4b50a628.