Move all Icinga contacts to monitoring::contacts
Contacts, priorities and time periods were previously split between the
`monitoring::checks` and `govuk::node::s_monitoring` classes. This made it
difficult to see how the three related to each other and muddied the actual
checks in `monitoring::checks`.

Improve this by consolidating them all in a new class and grouping together.
Change a couple of the param names to reflect what they are used for.
Document all of the params. Relocate the tests. This depends on
gds/deployment#441

It seems that we don't actually use the time periods. But I'll save removing
those until later, when I've double checked.