Disable centralised Syslog on Ubuntu 14.04
It appears that Ubuntu 14.04 is noticably more chatty on Syslog than
Ubuntu 12.04. We are also sending from a 7.4.4 Syslog to a 5.8.6 syslog.
At some point our syslog/logstash/elasticsearch infrastructure begins to
lag behind to the point where we are no longer indexing any syslog
entries in Elasticsearch.

This is a temporary fix to get syslog back for the majority of the
machines until we investigate further exactly what's causing us to fall
behind.