Install postgresql-contrib on slaves.
govuk_postgresql::extension is always defined as a virtual resource[1]
that's only realised on the non-slave servers[2].  This meant that the
postgresql-contrib package wasn't being installed on the slaves.  This
package provides the hstore extension (among others).  Without this,
pg_dump was failing when dumping a database that used hstore on the
slave[3].

This moves the include of postgresql::server::contrib to the server
package, which means it will be included on all postgresql servers.

[1] https://github.gds/gds/puppet/blob/master/modules/govuk_postgresql/manifests/db.pp#L80
[2] https://github.gds/gds/puppet/blob/master/modules/govuk_postgresql/manifests/server/not_slave.pp#L10
[3] https://deploy.production.alphagov.co.uk/job/Copy_Data_to_Staging/277/console