Expand the Hiera values for Graphite disk time.
The original code to boost these memory thresholds was introduced in
commit fad73d6ded4926493fac9b96214e7914e06f1794, but it doesn't appear
as though this is currently in effect in production. The Puppet
documentation[0] doesn't mention this nested property syntax. There are
therefore two possibilities:

 * this worked in an older version of Puppet/Hiera, but has been removed
 * this never worked, and I'm a numpty

The latter seems more likely.

Either way, we don't use this syntax for property overriding anywhere
else in our Hiera config (as far as I could find), even in situations
where it would make the code much more concise, so the expanded property
names definitely seem to be the standing convention, whether or not
that's because the other form just doesn't work.

[0]: https://docs.puppetlabs.com/hiera/1/puppet.html#automatic-parameter-lookup