[#81663690] Move assets canary content into Nginx
It occurred to me that if we have to configure the infrastructure to set
headers for this URL then we might as well put the content in the same
place, so that they're not unnecessarily decoupled.

This supersedes the static file in alphagov/static#476. It also changes the
response body to JSON to match the frontend/backend canary, but that
shouldn't affect the string check in Pingdom/Nimsoft.