Disregard negative replication lag in Postgres.
WARNING: this commit message contains unsubstantiated speculation.

Under some circumstances, it is possible for the existing replication
lag check to report a negative value, implying that the slave is further
ahead than its master. Possible explanations for this include:

  * the measurement being taken from the slave slightly later than from
    the master
  * umm... something something transactions?

This negative number is causing critical alerts in Icinga, despite the
best efforts of the `movingMedian` function: this may mean that the
negative number is being treated as a very large unsigned positive
number. Because a negative replication lag is not indicative of the
problem this check is looking for (namely that the slave is falling
behind) I think it's reasonable to drop these values.

I did consider using the `absolute` function to let us monitor this
information and alert if the slave got too far ahead, but decided
against it on the grounds that it would be confusing (and that, as far
as we know, a negative replication lag doesn't actually indicate a
problem).