Update mongo backend machines to use hieradata
The `configure_replica_set` class is now called via the `server` class. This commit passes the members of the replicaset to the server class from hiera to the new parameter, rather than calling the class directly (which will no longer work).

It is set in `class/mongo.yaml` rather than in the deployment project because the latter would require setting the same values in multiple places. This means there is a risk that it could be accidentally duplicated in hieradata. Discussed with @dancarley, and we feel that this risk is better than the risks inherent in duplicating the same code in several places.

This has been tested with vagrant-govuk and works. The first two members of the replicaset raise an error because the other members are not present, but the third one does not and at the end of `vagrant up` the replicaset is configured correctly, so I think these errors are acceptable.

In the process of moving, I have circumvented a legacy alias ('backend-1.mongo' => 'mongo-1.backend'). These legacy aliases may well be able to be removed now and I've added a FIXME.