Remove dupe exec for assets backup gpgkey
The duplicity module that we're using already takes care of downloading the
GPG pubkey for the key ID specified:

- https://github.com/alphagov/puppet-duplicity/blob/73db2dc81f49ffc5b4aeb0c99879f7ae05de9cc4/manifests/job.pp#L164-L168

I also don't think this was working correctly because the `gpgkey` variable
isn't declared. This just happened to be a noop because `--list-keys` (and
indeed `--recv-keys`) without any other arguments will always return a zero
exit code.