Remove ssh_id param from assets duplicity backups
SSH options, specifically IdentityFile, aren't passed by duplicity when
using rsync backends. As demonstrated by the arguments in the two commands
of this output:

    root@backup-1:~# bash -x /var/spool/duplicity/offsite-govuk-data.sh
    + set -e
    + set -o pipefail
    + duplicity --verbosity warning --no-print-statistics --full-if-older-than 30D --s3-use-new-style --gpg-options --trust-model=always --encrypt-key 13B84C37AB52D76B3F53CF0E7C34BD7A05119BA4 --ssh-optionsIdentityFile=/home/govuk-backup/.ssh/id_rsa --include '/data/backups/*/var/lib/automongodbbackup/latest' --include '/data/backups/*/var/lib/automysqlbackup/latest.tbz2' --include '/data/backups/*/var/lautopostgresqlbackup/latest.tbz2' --include /data/backups/archived --exclude '**' --archive-dir /root/.cache/duplicity/ / rsync://govuk-backup@172.16.10.10///srv/backup-data/
    Running 'rsync -e 'ssh -oBatchMode=yes' govuk-backup@172.16.10.10:/srv/backup-data/' failed with code 12 (attempt #1)

The bug is also described here:

- https://bugs.launchpad.net/duplicity/+bug/822697

The assets backup works because rsync and SSH use the default private key
location of `~/.ssh/id_rsa`. This happens to be the same as the path that
we're providing (the cron runs as root). So remove the explicit param, to
use defaults, and not make it appear like this functionality actually works.