[#80824874] Replace offsite-backup with duplicity
Use duplicity to do offsite backups instead of our custom script. This
performs GPG encryption and remote sync as before, but with the added
benefits that:

- we can re-use the same simple Puppet resource instead of duplicating the
  original shell script (this has copied `backup::assets`)
- it's less in-house code for us to maintain
- we save bandwidth and storage by using differential updates
- rotation of old backups is handled by the sender and not out-of-band by
  the receiver

WRT the last point, we can remove `base::rotate` from the
alphagov/govuk_offsitebackup-puppet once we've been running duplicity for
30days. All of the old style backups will have been removed and we'll have
the complete retention period for the new style.

We have lost the additional syslog logging and statsd timers because they
can't be implemented without making more changes to the duplicity module. We
should be able to get some of this information from duplicity's own logs.

The `enable` param is still supported but we don't setup the cronjob or the
alert if it is false (default). This too would have required significant
changes to the duplicity module to re-implement.

I've surfaced the GPG key (which can now be shared with `backup::assets`)
and the list of directories up to hiera, so that they are no longer hidden
within the logic. The directory list now just specifies the parent directory
for `archived` and `automongodbbackup` instead of a glob because those dirs
don't contain anything that we don't want.