Temporarily stop crawling /performance
When crawling /performance we crawl some PNG endpoints such as:

   /performance/site-activity-driver-and-vehicle-standards-agency/how-users-access-the-content/how-users-access-the-content-devices.png?selector=.visualisation-inner

These are routed through to the Performance Platform infrastructure and
are quite expensive to generate. Crawling them at the rate we crawl
(which is unconstrained unless we get a 429 response, but currently
1req/s) is introducing an unacceptable load on the platform.

This commit blacklists the entire of /performance while we investigate
better ways of dealing with this.