Set random_page_cost to 2.5 for transition-postgresql-*
This setting (default 4.0) is an estimate of the cost of a non-sequentially-fetched
disk page, in relation to seq_page_cost (which we're leaving at 1.0, the
default).

For Transition's largest hits queries, when random_page_cost is 4.0
the query planner decides to do a full table scan on hits, but when
it's set to 2.5 it's happier to use indexes. This makes these queries
faster, particularly in Preview, eg for HMRC all hits all time:

    random_page_cost       2.5     4.0
    Preview (Carrenza)    5.7s   14.5s
    Staging (Skyscape)    4.9s    7.4s