Set-up `real_ip_header` in node definition
`real_ip_header` is set in the deployment repo (see
https://github.gds/gds/deployment/blob/nginx-trueip/puppet/hieradata/production.yaml#L49).

Set `real_ip_header` up as a class parameter with a default value. If the
header doesn't exist, the value is set to undef so nothing is passed
through.

If it does exist, the value will be passed straight in to the Nginx
configuration because templates are rendered at runtime and class params are
available to them.