[#80824874] Send graphite backups offsite
Send onsite backups of Graphite data to the offsite machine. Almost the same
as the datastores offsite, except that these are unencrypted due to the
size, frequency of change, and low sensitivity of the data.

We have to use a separate duplicity job not to encrypt them. Which is
probably a good thing anyway, as we wouldn't want these large and frequently
changing files to muddy the quite small differential updates to the
datastore backups. We also need to double up some of the parameters that we
pass to this class.

I've set them both at the same time for want of finding a new cron time that
doesn't conflict with something else. They shouldn't interfere, at worst
they will both run slightly slower in total.