Default vhost_protected to false
The original logic inside `govuk::app::nginx_vhost` set the value of
`$protected_real` based on the environment it was being used it. To
accomplish this, we passed a messy chain of `undef` all the way down to
this class and made a logic switch. You can see the original behaviour
in https://github.gds/gds/puppet/blob/3bb03253c899ff261b7ce47f670e5d9ae0a78d8c/modules/govuk/manifests/app/nginx_vhost.pp#L27-L37

The existing befhaviour is that if `undef` is passed in, set it to
false. Otherwise, set it to the value passed in.

This changes the default all the way up the chain to be false, which
will allow a lot of simplification in calls to govuk::app which I will
make in a separate commit.