Add link to cron output for backup sync
The output of cron jobs is sent to the root email address on the machine
which is forwarded on to a google group.

If the backup sync failed, there may be useful diagnostic output from
cron. This link goes to a filtered view of the google group showing the
relevant emails.