Move the naming of the replicaset up a level
Since there is now a replicaset in dev too, it makes sense for that to be called 'development' not 'production'. This means both `mongodb.conf` and `configure-replica-set.js` need to have access to that variable, so this commit moves the decision up to the server class and then passes the replicaset name to both. (A subsequent commit will change the variable to a more useful name).

This commit also reformats a test that was modified for this behaviour. A subsequent commit will add another test for the behaviour.