Create per-app symlinks in /data/vhost directory
This means that the deploy scripts don't need to be reliant on the
app_domain.  They're currently constructing this path in a very brittle
way[1], which will also break should the app_domain ever be changed.
It's also necesitated some hacks for apps that deploy to multiple places
with differrent vhosts[2].

[1] https://github.gds/gds/alphagov-deployment/blob/master/recipes/defaults.rb (line 4 and 8)
[2] https://github.gds/gds/alphagov-deployment/blob/master/contacts/config/deploy.rb#L71-L83