Move backup::assets jobs to hiera
Follows the preceding commit for `backup::offsite`. I'm using YAML anchors
to de-dupe some of this data (that doesn't need interpolation) without the
need for additional hiera lookups (which are probably not very efficient).

It appears to be valid to have multiple keys called `_`. This namespacing
means that they'll never be used by Puppet. This commit was a noop on
`asset-slave-1` in Vagrant with `s_asset_slave::offsite_backups` enabled.