Move api-lb server config to env specific yaml files.
To make it consistent with the config for the other lb servers.

The content-store-3.api server is removed from the vagrant config
because this server doesn't exist in preview, and therefore won't exist
in vagrant (vagrant-govuk looks at the preview config in the
provisioning repo for the list of servers to create).