Receive Bouncer's CDN logs on logs-cdn-1
These logs provide crucial data for the Transition app: we process the logs and
import new data from them hourly into the app, and our users rely on the data
to show them where mappings have been missed, or may not be meeting user needs.

At the moment we stream these logs to transition-logs-1 in the CI environment,
and process them there. Over the last few months the streaming has broken
multiple times, often over the weekend for several days at a time, resulting in
data loss and obvious gaps in the graphs in the Transition app. This reduces
users' confidence in the data which *is* there, as well as reducing the
information available to them.

We want to set up monitoring for the log streaming so that 2nd line is notified
if it breaks. Moving the streaming to logs-cdn-1 in Production brings it in
line with other existing CDN logs, and allows us to use Nagios checks, whereas
monitoring them in the CI environment would mean setting up non-standard email
alerts, which would be another source of alerts to keep track of and make it
more likely that those alerts would be ignored or forgotten about.

We will set up an additional logstream to logs-cdn-1, and when we're confident
it is working as intended we will start sending the logfiles from there to
transition-logs-1 for processing every hour and turn off the existing stream
to that machine.

We don't know why the logs streaming breaks, but restarting rsyslog has fixed
it every time so far. We would like to do this automatically, but will tackle
that in a later PR.