Stop enabling pgcrypto extension for Transition's database
We needed this extension to calculate hashes of paths during Transition's
hits import, as a hangover from MySQL. We don't need to do that any more
now that we're on Postgres, and have stopped doing it, so don't need
this extension any more.

We will devops the databases in all environments after this has been
deployed to drop the extension for this database - we don't think
there's a need to automate that part of the process in Puppet.

The plpgsql extension is enabled by default for all databases, so
we won't drop that one, but don't need to specify it here either.