Fix protection for frontend on backends
Whitehall Frontend and Frontend a "special cases" that were missed in
the frontend and the backend, however on the backend we want them to be
protected but on the frontend we do not.

This means that we can't remove vhost_protected from the govuk::apps
class, although we can default it to false and only enable it on backend
servers.