[#83054216] Revert mongodb upstart to pkg default
When we started using mongodb we were using odd packages that did not
provide an upstart script, so we copypasted our own from an unknown
source.

This upstart script sets various options, but most importantly contains
a piece of logic to do automated "repairs" which we don't have a clear
use case for an actually presents a significant issue. When mongodb is
shut down cleanly it leaves an _empty_ pidfile. The repair function
assumes that the existence of a pidfile implies an unclean shutdown so
it _always_ runs a repair which can take upwards of 40 minutes on some
machines.

The mongodb packages we currently use would normally provide their own
upstart script which has the following differences from ours:

1. Slightly lower nofile limits (20000 vs 32000)
2. No nproc ulimit so defaults to whatever limits the binary requests.
   - In my testing the maximum number of threads I've found on a mongodb
     server is 190, whereas the nproc that the process requests is 2741.
     Should we need to set a ulimit later, it can be done in
     /etc/defaults/mongodb
3. No repair functionality
4. No special logging options so the default logging of stdout captured by
   upstart to /var/log/upstart/mongodb.log
5. Uses `--exec` vs. `--startas` (no practical difference in this context)

So this commit:

1. Removes our special logging configuration
2. Replaces our upstart script with the one from the currently installed
   package (mongodb-10gen v 2.4.9) with appropriate FIXME for later to
   stop managing this file, at which point it will revert to being
   managed by the package.