[#80824874] Add another disk to backup-1
This will be used to grow the existing logical volume for onsite backups so
that we can transfer the Graphite data uncompressed.

Accompanies gds/govuk-provisioning#166 which has been applied manually to
Preview (because vcloud-launcher doesn't handle additional disks). I'll do
the same for Staging and Production tomorrow.

I've tested this in Vagrant to confirm that `govuk::lvm` will create a new
physical volume and add it to the volume group, but it won't extend the
logical volume or the filesystem because we don't specify a `size` param.
Which is probably a good thing - I'll do that part by hand and update the
opsmanual accordingly.