[#80824874] Don't tarball Graphite backups
Duplicity on `backup-1.management` was taking more than 24hrs to send
incremental backups of the Graphite tarball. It seemed that a lot of this
time was spent computing the differences, rather than actually sending any
data.

There are some very anecdotal references to Duplicity not dealing with large
files well and a suggestion that it might have improved in later versions
(that aren't available to us on Ubuntu LTS):

- https://bugs.launchpad.net/duplicity/+bug/1301892
- https://bugs.launchpad.net/duplicity/+bug/897423

I performed a slightly crude test whereby I rsync-ed 15G of whisper data
from `graphite-1` to `backup-1` and then ran duplicity to send it to the
off-site machine which took 18mins each and resulted in 3G of off-site data.
I then repeated both steps, which took 5mins each and resulted <0.5G of
additional off-site data.

This isn't as much data as we normally backup - both in terms of total
quantity and differential changes. But I think it demonstrates that it's
more efficient to off-site the data uncompressed than compressed. I believe
this is because rsync and duplicity are able to detect that only the first N
bytes of each whisper file has changed. Whereas the same data compressed has
many more adjacent changes.

This will unfortunately use more disk space on the on-site machine as added
in d0543e4. However it will save space on the off-site machine, plus time
and CPU cycles throughout. It also makes this on-site job look like all of
our other on-site backups.

The old tarball will be cleaned up on the on-site machine by the `rsync
--delete` in `backup::directory`. However we'll need to clean up the old
duplcity backup on the off-site machine immediately after deploying this
change.