[#47635423] Don't manage file perms for Nginx logs
The `nginx::log` resource was originally created prior to logstream/tagalog
to set ownership and permissions on files in `/var/log/nginx`. This was for
the story [#36626383] because it was noted that the ownership and
permissions on existing machines was inconsistent.

The reason for that inconsistency is that when Nginx first starts it is the
"master process", which runs as `root:root`, that initially opens the log
files:

    root@graphite-1:~# rm /var/log/nginx/{access,error}.log
    root@graphite-1:~# service nginx restart
     * Restarting nginx nginx
       ...done.
    root@graphite-1:~# ls -al /var/log/nginx/{access,error}.log
    -rw-r--r-- 1 root root 0 Dec  4 13:53 /var/log/nginx/access.log
    -rw-r--r-- 1 root root 0 Dec  4 13:53 /var/log/nginx/error.log

When the files are rotated by logrotate they get recreated as `www-data:adm`
so that the privsep "worker processes", which also run as that user, are
able to write to them:

    root@graphite-1:~# grep -E 'create|kill' /etc/logrotate.d/nginx
      create 0640 www-data adm
        [ ! -f /var/run/nginx.pid ] || kill -USR1 `cat /var/run/nginx.pid`

This matches with Nginx's behaviour when you issue a `USR1` signal. It
changes the ownership of the files from `root` to `www-data`:

    root@graphite-1:~# ls -al /var/log/nginx/{access,error}.log
    -rw-r--r-- 1 root root 0 Dec  4 13:53 /var/log/nginx/access.log
    -rw-r--r-- 1 root root 0 Dec  4 13:53 /var/log/nginx/error.log
    root@graphite-1:~# pkill -USR1 nginx
    root@graphite-1:~# ls -al /var/log/nginx/{access,error}.log
    -rw-r--r-- 1 www-data root 0 Dec  4 13:53 /var/log/nginx/access.log
    -rw-r--r-- 1 www-data root 0 Dec  4 13:53 /var/log/nginx/error.log

This inconsistency doesn't strictly matter. The only other thing that reads
from these files is logstream/tagalog (soon to be replaced) which also runs
as `root` so that it can read all log files.

So stop managing these in favour of simplicity and faster provisioning times
(less Puppet resources to execute). We can remove all `nginx::log` resources
that only specified `logstream => absent` (predating logstream) because
these now have no effect.

There is a small "regression", back to the standard Nginx/Ubuntu behaviour,
that the initial log files will be world readable until their first
rotation. I don't think this matters though - we trust people that have
shell access.