Defaults for rspec-puppet global hiera keys
Set some default data for global hiera keys that are indirectly referenced
by many classes. This is similar to what we do for `default_facts` and saves
us from muddying all of our tests with unrelated setup.

The remaining instances of `let(:hiera_data)` are providing values that are
specific to the tests in that context, e.g. if `app_domain` is `test.gov.uk`
then we should have a vhost of `foo.test.gov.uk`.

The `SharedContext` overrides as similar setup in hiera-puppet-helper:

https://github.com/bobtfish/hiera-puppet-helper/blob/2853187c6553ee292793292c852a9e57c58da167/lib/hiera-puppet-helper/rspec.rb#L3-L11