Remove rubygems fixme
The origin of these three resources was to fix a problem caused by us
installing a non-Ubuntu supplied system ruby with rubygems in a weird
location. The resources removed that location for gems and reverted to
the system defaults.

The resources contain a problem that manifests on new machines:

```
  exec { 'migrate /usr/lib/ruby/gems to /var/lib/gems':
    user => 'root',
    command => 'mv /usr/lib/ruby/gems /var/lib/',
    unless => 'test -d /var/lib/gems',
    require => Exec['reinstall correct ruby1.9.1 package'],
  }
```

The unless test runs and on a machine where /var/lib/gems/ exists it will
exit with 0 and the `command` will not run. On a machine where
/var/lib/gems doesn't exist (as on a new machine) it's going to exit
with a 1, which means that the `command` will run. However
/usr/lib/ruby/gems doesn't exist on a new machine either, so the `mv` will
fail.

I did consider fixing this by adding conditions to check if both
/var/lib/gems and /usr/lib/ruby/gems exist, but in the end I've simply
checked that we don't currently install the non-Ubuntu rubygems version
currently and all machines have been fixed and am suggesting removing
the FIXME.