Add class to install openjdk7
In order to support using openjdk rather than java provided by Oracle,
we need a class which installs the correct packages and the ability to
set that version to be the default.

In order to support installing openjdk-7 I have copied the openjdk-6
class. I checked and it would have been possible to use openjdk6 for
what I am about to do (packages exist for Precise and Trusty) but I feel
we should target a more modern version.