Switch govuk_jenkins to openjdk7
In theory, the jenkins packages we use should run perfectly well with
openjdk. In order to reduce our reliance on the Oracle java packages
which we install and maintain (with special code to accept the licence)
we should switch as many things as possible away from them.

This:

1. Removes the oracle7 packages
2. Installs the openjdk7 packages
3. Switches the default `java` binary to be openjdk
4. Installs the Github Enterprise cert to the right keystore

I've tested that the result of switching to this branch is that the
jenkins war file is running under `/usr/bin/java` and this is the
openjdk7 provided version.