Add aptly::repo for Heka
Heka don't release their packages to an APT repo. Instead they are available
as `.deb` files from the GitHub releases page:

https://github.com/mozilla-services/heka/releases

As such there isn't anything that we can mirror like we do for other
third-party packages and the packages are of good enough quality that it
doesn't warrant attempting/forcing a rebuild through our PPA process.

This uses the `aptly::repo` defined type (contributed by IDA) to create an
empty repo. Afterwards I'll need to run:

    aptly repo add heka heka_0.8.0_amd64.deb
    aptly snapshot create heka-$(date +%Y%m%d) from repo heka
    aptly publish snapshot heka-$(date +%Y%m%d)

Not sure at this stage whether this is worth adding to the opsmanual.
Hopefully we shouldn't have to do this too often.