Namespace hiera deploy_ssh_keys
This hiera key is only used by a single class so it should be namespaced
accordingly. This removes one of our unnecessary uses of `let(:hiera_data)`
and a whitelist entry from the lint check.

Also changed matcher in test from `should_not … with_content` to `should …
without_content` because it's more reliable as described in 09895cc.

Depends on gds/deployment#464