Add govuk_etcd class to configure an etcd cluster
Add a new class `govuk_etcd`, that acts as a thin wrapper for the
upstream KyleAnderson/etcd Puppet module.

Note that we configure the first node in the cluster with an empty list
of peers to allow the cluster to bootstrap when it first comes up.

If we try to pass the peer list to all nodes, the cluster fails to start
as each node waits until it can connect to a peer before it starts
listening on port 7001 (which is used for peer communication).

By starting the one node with no peers, it promotes itself to the
leader and starts listening on port 7001. When the other nodes come up,
they are able to connect to the first node and start participating in
the cluster.

I have tested this in Vagrant; I tried destroying the data on one node
then bringing that node back online and the cluster re-synchronised the
data.