Correct expected path of openjdk binaries
In order to determine whether set_defaults needs to make changes it
reads the destination of the /etc/alternatives/{java,javac} symlinks to
determine if they match expectations.

The correct paths when openjdk7 is configured are:

```
$ readlink /etc/alternatives/java
/usr/lib/jvm/java-7-openjdk-amd64/jre/bin/java
$ readlink /etc/alternatives/javac
/usr/lib/jvm/java-7-openjdk-amd64/bin/javac
```

Because the current paths are missing `-amd64` the `unless` condition
does not match and the `set_defaults` class is *always* refreshed. This
wouldn't be so bad, but it notifies `jenkins::service` which means every
time Puppet runs on the Jenkins machines, Jenkins restarts - roughly on
the hour and on the half-hour...

So this commit corrects the path to the `javac` and `java` binaries
provided by openjdk7.