Remove /healthcheck endpoint from default vhost
Having a `/healthcheck` endpoint in the default Nginx virtual host is
potentially dangerous as it could be construed as an application
healthcheck (which use the same endpoint) and lead someone (or a
monitoring application) to believe that an application is healthy even
if the application has not yet been deployed.

I considered renaming the endpoint to `/default_vhost_healthcheck` to
make its provenance clear, however after some investigation this
endpoint is not used by any checks. I'm therefore removing it to
simplify our configuration.

The load balancers for Bouncer and EFG [1] use the `/healthcheck` endpoint
without specifying a hostname which I thought would be served by the
default virtual host, however those machines do not use a default
virtual host.

[1]: https://github.gds/gds/govuk-provisioning/blob/5f65025ff4e0995c496ce3f53c64dba343896918/vcloud-edge_gateway/rules/lb.yaml.mustache#L16-L34