Change varnish alert to use Kibana for action_url
I'm going to be changing the names of the statsd metrics used in this
Graphite URL as part of the "remove tagalog" work, so this will no longer
work.

Instead of updating it to the new names I'm changing it for a Kibana URL
which returns the log entries for HTTP responses that had 5xx status codes.
I've also reduce the time period from 8hrs to 4hrs because it's easier to
see the most recent alerts. This should provide more context for someone on
2nd line responding to this alert.

No tests because this is an exported resource :(