Heka TCP input + Elasticsearch output on s_logging
Use `s_logging` as a centralisation point for Heka logs. Technically each
node could output to Elasticsearch directly, but we'd need to open up
firewall access for this to happen and it would expose the ability for a
rogue machine to delete/modify existing data in ES.

The `logs-elasticsearch.cluster` hostname being used here is, despite it's
name, not a cluster. It points at a single host and is a single point of
failure. To be addressed in a following commit.