Use Elasticsearch local proxy for Heka output
So that one Elasticsearch node doesn't become a single point of failure.
Unfortunately this introduces Nginx to this machine. But I think that's a
reasonable trade-off.

We tried using this for Logstash in the past but reverted it in 6496b5a.
This appears to work fine, in testing, for Heka.