Heka on all machines forwarding to s_logging
Setup Heka on all machines and make them forward logs to `s_logging`.

This will be used to replace tagalog/logstream, which runs on *all* machines
currently. Local inputs will be configured in subsequent commits.

TCP output has been disabled on `s_logging` because we don't want it to
forward to itself.