Fix Heka on the development VM
Modules like Nginx, that include `heka` classes, will install and configure
Heka on the development VM. This is undesirable because we don't care for
shipping logs from this VM - per 27895d9.

Make the repo available, as we do in `s_base`, so that it can be installed
and configured. We can't easily prevent this from happening. But we can then
ensure that the service isn't running so it's not consuming any CPU and RAM.