Consume all Nginx error logs with Heka
Read all Nginx error log files using Heka. Includes the default log and all
other vhosts. By globbing for these instead of using individual Puppet
resources we:

- guarantee that we won't forget any
- speed up Puppet by reducing the number of resources

Nginx doesn't give us any metadata error logs, like `server_name` or
`http_host`, so these will be identifiable in Kibana by the basename of the
logfile which goes into the `@logger` field - that's what the
`differentiator` option is doing.

Using Whitehall as an example where the log basename matches `server_name`:

    @type:nginx.error AND @logger:whitehall-frontend
    @type:nginx.error AND @logger:whitehall-frontend*
    @type:nginx.error AND @logger:"whitehall-frontend.preview.alphagov.co.uk"

But some of the others are more abstract:

    @type:nginx.error AND @logger:lb
    @type:nginx.error AND @logger:graphite
    @type:nginx.error AND @logger:default

We have to set the `hostname` option to ensure that `@source_host` contains
the FQDN instead of the unqualified hostname. This is because Go's
`os.Hostname()`, which is what Heka defaults to, only returns the short name
on our machines. In future versions of Heka this will be a global option so
we don't need to configure it for every log file.