Consume all Nginx access logs with Heka
This is basically the same as errors logs two commit previous. The only real
difference is that we're using the JSON event decoder and they have
`@type:nginx.access`.

These entries can be searched in Kibana using `server_name:` and
`http_host:` as before. But you may wish to use `@logger:` instead if you
want to see access and error logs for the same vhost.