Heka emit statsd metrics for Nginx access logs
This reproduces a feature of tagalog/logship (and prior to that logster).
It has some differences with regards to naming though:

I'm using `@logger` for the "vhost name". This is set to be the same as the
basename of the log file (minus the `.json.event.access.log` suffix) by
Heka's `differentiator` option.

I chose this because it's the most consistent variable that we have and
means that we can be sure that the statsd key that we're checking with
`icinga::check::graphite` is the same thing that we're emitting in Heka. It
is also consistent with error logs which means that you can search for
access and error logs for the same thing in Kibana using `@type:nginx.* AND
@logger:<vhost>`.

Some other things that I considered were:

- `@fields.server_name`, but we don't set this consistently. Some vhosts
  have wildcards in the primary name, are automatically generated, or have
  no value at all.
- `@fields.http_host` is even more inconsistent because it's whichever
  expanded `server_name` entry the client matched using a `Host` header.
  Likewise some requests have no value at all.
- A custom Nginx variable added to `log_format json_event` and populated by
  `set` in each vhost. This seemed very brittle because we'd need to
  remember to set it in all new vhosts. It would also clutter the log
  format/storage and required increasing Nginx's `variables_hash_max_size`
  option.
- A value modified or inserted using a chained encoder (filters can create
  but not modify) in Heka. But this would be complicated, decouple the
  naming from Puppet, and potentially still result in storing the additional
  field in Elasticsearch.

Because we can't easily modify the value (without changing our log naming
scheme, which seems like bad coupling) we have also regressed on one of the
changes in a2f2273 which prevented deep namespacing in statsd. Instead of
`whitehall-admin_preview_alphagov_co_uk` we will have
`whitehall-admin.preview.alphagov.co.uk` where each dot represents a level
in the hierarchy. I'm not that keen on this, but it has no functional effect
and we don't look through that namespace by hand very often.

A lot of these problems could be solved by using individual Puppet resources
to create Heka config snippets that monitor individual files and set the
appropriate values on each. The benefit of not doing that though is that we
speed up Puppet by managing less resources, decrease complexity in our
Puppet code, and don't risk forgetting to monitor new log files - which I
think makes these things a worthwhile trade-off.

We're no longer using statsd's legacy namespacing. So counters will be
prefixed with `stats.counters.<name>` instead of `stats.<name>` and suffixed
with `<name>.rate` where previously there was no suffix at all:

- http://hekad.readthedocs.org/en/v0.8.0/config/inputs/index.html#stataccuminput
- https://github.com/etsy/statsd/blob/master/docs/namespacing.md

The matcher on the `StatFilter` prevents us from generating stats when we
don't have any valid data to submit. The `request_time` and `status` checks
could be split out to separate `StatFilter` objects, but for simplicity I'm
just ignoring entries that don't have both.

The `ticker_interval` on `StatAccumInput` matches how we currently
configured statsd to submit metrics to Graphite.